Additional progress
process.py -- SHOULD be feature complete. Testing needs to be done. I want to add a KeyboardInterrupt exception in the future

mapview.py -- (WIP) Due to modern browsers and CORS restrictions, viewing the existing index.html locally will not work since it tries to access our hasGig.csv (in the future addresses.db) file and is restricted. So this needs to be web server hosted. This process runs a barebones local HTTP server to accomplish this.