Fixes
process.py -- Changed selection query to only pull addresses from the database that have the lightGig flag of 0 or NULL to only process addresses that we have not confirmed to have fiber. This is to lessen the polling to AT&T and thus processing time. If anyone has a reason we should continue to check addresses that are already verified, please let me know.

openaddresses-selection.py -- Changed lat/lon order to match the order given by Google Maps when copying coords to save a small bit of work on the user end.