Update process.py
Small tweaks. If the check with ATT fails, we break and (hopefully) go to end of script and properly commit/close the DB file. It's assumed in most cases if the ATT check fails, they may have blocked us and will continue in the interim.

Added debug lines to the DB commit and close lines for visual confirmation when this step has been reached. Have had no shortage of issues so far getting these to take place when something else breaks so this should help with testing.